backend/requirements.txt
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
graphql-core==3.2.5
gyp==0.1
itsdangerous==2.2.0
Jinja2==3.1.5
mariadb==1.1.11
MarkupSafe==3.0.2
packaging==24.2
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
strawberry-graphql==0.258.0
typing_extensions==4.12.2
Werkzeug==3.1.3